requirements.txt
opencv-python>=4.5.0
numpy>=1.20.0
scipy>=1.7.0
customtkinter>=5.0.0
Pillow>=9.0.0
comtypes>=1.1.0
hidapi>=0.14.0
